RF: moved test_json_dumps_pretty into test_utils
